dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=dev-requirements.txt dev-requirements.in setup.py
#
attrs==19.3.0             # via pytest
blinker==1.4
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask, pip-tools
entrypoints==0.3          # via flake8
faker==4.0.1
flake8==3.7.9
flask==1.1.1
idna==2.8                 # via requests
importlib-metadata==1.5.0  # via pluggy, pytest
itsdangerous==1.1.0       # via flask
jinja2==2.11.1            # via flask
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==8.2.0     # via pytest
packaging==20.1           # via pytest
pip-tools==4.4.1
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pyparsing==2.4.6          # via packaging
pytest==5.3.5
python-dateutil==2.6.1    # via faker
python-dotenv==0.11.0
requests==2.22.0
responses==0.10.9
six==1.14.0               # via packaging, pip-tools, python-dateutil, responses
text-unidecode==1.3       # via faker
urllib3==1.25.8           # via requests
wcwidth==0.1.8            # via pytest
werkzeug==1.0.0           # via flask
zipp==2.2.0               # via importlib-metadata
